valamiert ha uint64on vannak a timer szamai akkor a egyes animaciok lehalnak still, timer is next

+++ SDLProject/include/Timer.hpp
@@ -10,8 +10,8 @@ class Timer
 private:
 	Timer() {};
 	static Timer* Timer_Instance;
-	Uint64 Timer_DeltaTime;
-	Uint64 Timer_LastTime;
+	double Timer_DeltaTime;
+	double Timer_LastTime;
 public:
 	inline static Timer* GetInstance()
 	{
